refactor: comprehensive issue swarm remediation (#1995, #1997, #2000, #2002, #2003)
Address 5 Tools issues via parallel agent swarm in a single coordinated PR.

## Issue #2000: Decompose kalman_filter.filter() (90 LOC)
- src/data_processing/data_processor/python/data_processor/core/kalman_filter.py
- Extract predict/update/accumulate steps into dedicated private methods
- filter() becomes a thin orchestrator
- Preserves exact public API and numerical behavior

## Issue #1995: Decompose pendulum-core/src/lib.rs (1618 LOC)
- src/pendulum_simulator/pendulum-core/src/lib.rs
- Split into focused modules under src/bindings/:
  - state.rs, dynamics.rs, integrator.rs, forces.rs, energy.rs, mod.rs
- lib.rs now a thin public API + re-exports
- Preserves all public APIs (no breaking changes)

## Issue #2002: Decompose top 5 oversized Python functions
- src/pendulum_simulator/src/double_pendulum_golf/gui/toolstrip_widget.py
- src/pendulum_simulator/src/double_pendulum_golf/gui/panel_builders.py
- src/pendulum_simulator/src/double_pendulum_golf/gui/analysis_tab.py
- src/pendulum_simulator/src/double_pendulum_golf/gui/optimization_widget.py
- Extract long build/setup methods into private helper methods <40 LOC each
- Each orchestrator stays <50 LOC

## Issue #2003: Split top 3 monolithic scripts
- src/solar_system_model/solar_system/data/historical_events.py (920 LOC)
  -> split into historical_events/ package with ancient_and_early, crewed_programs,
     international_missions, modern_era, etc. (each <400 LOC)
- src/solar_system_model/solar_system/ui/widgets.py (1022 LOC)
  -> split into widgets/ package with settings_nav, sidebar_controls, etc.
- src/pendulum_simulator/src/double_pendulum_golf/gui/simulation_panel.py
  -> split into simulation_panel/ package with _simulation_panel, _lifecycle_mixin,
     _export_mixin, _worker modules
- All __init__.py files provide re-exports so existing imports keep working

## Issue #1997: 500 LOC CI gate
- scripts/check_file_size_budget.py: Python CLI for file-size enforcement
  - Supports --max-loc, --changed-only, --baseline-file flags
  - Uses logging (no print calls)
  - <200 LOC, type-hinted, pathlib-based
- scripts/monolith_baseline.txt: grandfathered list of ~171 existing monoliths
- docs/assessments/monolith_inventory_2026-04-10.md: inventory doc with burn-down plan
- .github/workflows/file-size-budget.yml: CI workflow running the gate on PRs
- src/shared/tools/README.md: conventions for cross-tool helpers

## Stale issues closed (companion to this PR)
- #2004: stale (0 real print() violations, per ruff T201 check)
- #1996, #1999: stale (equations_data.py already removed)

## Verification
- ruff check .: passes (5 line-length errors fixed in historical_events split)
- ruff format --check .: 1131 files already formatted
- pytest: 4522 passed, 21 skipped, 10 pre-existing failures (all from missing
  optional deps: sympy, httpx, fastapi; verified unchanged on staging baseline)
- Zero regressions introduced by this refactor
- All public APIs preserved
