+++ src/solar_system_model/solar_system/data/historical_events/ancient_and_early.py
@@ -1,0 +1,131 @@
+"""
+Ancient Astronomy and Early Space Age
+======================================
+
+Events spanning pre-1960s astronomy (Galileo onward) and the early space age
+(1957 Sputnik through the beginning of the 1970s), including the Apollo
+program.
+"""
+
+from typing import Any
+
+ANCIENT_AND_EARLY_EVENTS: list[dict[str, Any]] = [
+    # Ancient Astronomy
+    {
+        "year": 1610,
+        "month": 1,
+        "day": 7,
+        "title": "Galileo Discovers Jupiter's Moons",
+        "description": (
+            "Galileo Galilei observes four moons orbiting Jupiter, providing evidence"
+            " for Copernican heliocentrism"
+        ),
+        "category": "discovery",
+    },
+    # Early Space Age
+    {
+        "year": 1957,
+        "month": 10,
+        "day": 4,
+        "title": "Sputnik 1 Launched",
+        "description": (
+            "Soviet Union launches first artificial satellite, beginning the Space Age"
+        ),
+        "category": "mission",
+    },
+    {
+        "year": 1961,
+        "month": 4,
+        "day": 12,
+        "title": "First Human in Space",
+        "description": (
+            "Yuri Gagarin becomes the first human to orbit Earth aboard Vostok 1"
+        ),
+        "category": "mission",
+    },
+    # Apollo Program
+    {
+        "year": 1968,
+        "month": 12,
+        "day": 21,
+        "title": "Apollo 8 Launch",
+        "description": (
+            "First crewed mission to orbit the Moon, capturing the famous 'Earthrise'"
+            " photo"
+        ),
+        "category": "mission",
+    },
+    {
+        "year": 1969,
+        "month": 7,
+        "day": 16,
+        "title": "Apollo 11 Launch",
+        "description": (
+            "Saturn V rocket launches with Neil Armstrong, Buzz Aldrin, and Michael"
+            " Collins"
+        ),
+        "category": "mission",
+    },
+    {
+        "year": 1969,
+        "month": 7,
+        "day": 20,
+        "title": "First Moon Landing",
+        "description": (
+            "Apollo 11 lands in the Sea of Tranquility. 'That's one small step for"
+            " man...'"
+        ),
+        "category": "mission",
+    },
+    {
+        "year": 1969,
+        "month": 7,
+        "day": 21,
+        "title": "First Moonwalk",
+        "description": (
+            "Neil Armstrong and Buzz Aldrin walk on the lunar surface for 2.5 hours"
+        ),
+        "category": "mission",
+    },
+    {
+        "year": 1969,
+        "month": 7,
+        "day": 24,
+        "title": "Apollo 11 Returns",
+        "description": (
+            "Safe splashdown in Pacific Ocean, completing historic moon mission"
+        ),
+        "category": "mission",
+    },
+    {
+        "year": 1970,
+        "month": 4,
+        "day": 11,
+        "title": "Apollo 13 Launch",
+        "description": (
+            "Launch of Apollo 13, which would face a critical in-flight emergency"
+        ),
+        "category": "mission",
+    },
+    {
+        "year": 1970,
+        "month": 4,
+        "day": 13,
+        "title": "Apollo 13 Accident",
+        "description": (
+            "'Houston, we've had a problem' - oxygen tank explosion forces mission"
+            " abort"
+        ),
+        "category": "mission",
+    },
+    {
+        "year": 1970,
+        "month": 4,
+        "day": 17,
+        "title": "Apollo 13 Safe Return",
+        "description": (
+            "Crew safely returns to Earth after using lunar module as 'lifeboat'"
+        ),
+        "category": "mission",
+    },
+]
